refactor: digit 컴포넌트 useDigit hook으로 분리 및 버튼 요소 배열 수정
