+++ src/hooks/useOperatin.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+const useOperatin = () => {
+  return;
+};
+
+export default useOperatin;
