+++ src/types/allProps.d.ts
@@ -11,3 +11,5 @@ export interface IModifiersProps {
 export interface IOperationsProps extends ICalculator {}
 
 export interface ITotalProps extends Omit<ICalculator, "setCalculation"> {}
+
+export interface IUseDigtProps extends ICalculator {}
